+++ Common/V.cs
@@ -1,7 +1,6 @@
 namespace Common;
 
-public record V(int X, int Y)
-{
+public record V(long X, long Y) {
     public static readonly V Left = new(-1, 0);
     public static readonly V Right = new(1, 0);
     public static readonly V Down = new(0, 1);
@@ -16,37 +15,31 @@ public record V(int X, int Y)
     public static readonly V SE = S + E;
     public static readonly V SW = S + W;
 
-    public bool IsInRange<T>(T[][] map)
-    {
+    public bool IsInRange<T>(T[][] map) {
         return X >= 0 && X < map[0].Length && Y >= 0 && Y < map.Length;
     }
-    
-    public bool IsInRange<T>(Map<T> map)
-    {
+
+    public bool IsInRange<T>(Map<T> map) {
         return X >= 0 && X < map.SizeX && Y >= 0 && Y < map.SizeY;
     }
 
-    public bool IsInRange<T>(T[,] map)
-    {
+    public bool IsInRange<T>(T[,] map) {
         return X >= 0 && X < map.GetLength(1) && Y >= 0 && Y < map.GetLength(0);
     }
 
-    public static V Parse(string s)
-    {
+    public static V Parse(string s) {
         var tokens = s.Split(new[] { ',', ' ' }, 2);
         return new V(int.Parse(tokens[0]), int.Parse(tokens[1]));
     }
 
-    public static V[] Directions4 => new[]
-    {
+    public static V[] Directions4 => new[] {
         Left,
         Up,
         Right,
         Down,
     };
 
-    public static V[] Directions5 => new[]
-    {
+    public static V[] Directions5 => new[] {
         Left,
         Up,
         Right,
@@ -58,23 +51,21 @@ public record V(int X, int Y)
     public IEnumerable<V> Area4() => Directions4.Select(dir => this + dir);
     public IEnumerable<V> Area8() => Directions8.Select(dir => this + dir);
 
-    public static V[] Directions8 => new[]
-    {
+    public static V[] Directions8 => new[] {
         E, NE, N, NW, W, SW, S, SE
     };
 
     public static V Zero => new(0, 0);
 
-    public int DistTo(V other) => Math.Abs(other.X - X) + Math.Abs(other.Y - Y);
+    public long DistTo(V other) => Math.Abs(other.X - X) + Math.Abs(other.Y - Y);
     public V Abs() => new(Math.Abs(X), Math.Abs(Y));
 
-    public int CDistTo(V other) => Math.Max(Math.Abs(other.X - X), Math.Abs(other.Y - Y));
+    public long CDistTo(V other) => Math.Max(Math.Abs(other.X - X), Math.Abs(other.Y - Y));
 
-    public int MLen => Math.Abs(X) + Math.Abs(Y);
-    public int CLen => Math.Max(Math.Abs(X), Math.Abs(Y));
+    public long MLen => Math.Abs(X) + Math.Abs(Y);
+    public long CLen => Math.Max(Math.Abs(X), Math.Abs(Y));
 
-    public IEnumerable<V> LineTo(V other)
-    {
+    public IEnumerable<V> LineTo(V other) {
         var dir = (other - this).Signum();
         for (var point = this; point != other; point += dir)
             yield return point;
@@ -82,8 +73,7 @@ public record V(int X, int Y)
         yield return other;
     }
 
-    public V Rotate(int degrees)
-    {
+    public V Rotate(int degrees) {
         double angle = Math.PI * degrees / 180.0;
         (double sinAngle, double cosAngle) = Math.SinCos(angle);
 
@@ -93,7 +83,7 @@ public record V(int X, int Y)
         var y = (int)(X * sinAngle + Y * cosAngle);
         return new(x, y);
     }
-    
+
     public V RotateAround(int degrees, V pivot) {
         var translatedThis = this - pivot;
         return translatedThis.Rotate(degrees) + pivot;
@@ -111,5 +101,5 @@ public record V(int X, int Y)
 
     public V Signum() => new(Math.Sign(X), Math.Sign(Y));
 
-    public V Mod(int height, int width) => new V(X.Mod(width), Y.Mod(height));
+    public V Mod(long height, long width) => new V(X.Mod(width), Y.Mod(height));
 }
